requirements.txt
aiocache==0.12.1
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
easy-pil==0.3.0
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
langdetect==1.0.9
multidict==6.0.4
orjson==3.9.1
py-cord==2.4.1
pycares==4.3.0
pycparser==2.21
PyNaCl==1.5.0
pyreadline3==3.4.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2